requirements.txt
google-generativeai
faiss-cpu
sentence-transformers